Refactor GameEngine to use Environment.Exit for application termination instead of GetTree().Quit. Update AdvancedMagicShopLocation and AnchorRoadLocation to handle input more robustly. Rename finalSellPrice to calculatedSellPrice for clarity in item transactions across multiple locations. Introduce parameterless constructors in PrisonLocation and PrisonWalkLocation for compatibility. These changes enhance code clarity and improve input handling.

+++ Scripts/Core/Monster.cs
@@ -325,7 +325,7 @@ public class Monster
         {
             case 0: // Heal
                 var healAmount = MagicLevel * 10 + GD.RandRange(5, 15);
-                HP = Math.Min(HP + healAmount, GetMaxHP());
+                HP = Math.Min((int)HP + (int)healAmount, (int)GetMaxHP());
                 result.Message = $"{Name} heals for {healAmount} points!";
                 break;
